change interest incurred from double to big decimal

+++ src/main/java/africa/nkwadoma/nkwadoma/infrastructure/adapters/output/persistence/entity/loanentity/LoaneeLoanDetailEntity.java
@@ -24,5 +24,5 @@ public class LoaneeLoanDetailEntity {
     private BigDecimal amountRepaid;
     private BigDecimal amountOutstanding;
     private double interestRate;
-    private double interestIncurred;
+    private BigDecimal interestIncurred;
 }
